fetch forecast data from view model in fragment

+++ app/src/main/java/com/alalmiyaalhura/forecastapp/data/model/Forecast.kt
@@ -38,7 +38,7 @@ data class Forecast(
 
 @SerializedName("pop")
 @Expose
- val pop:Int? = null,
+ val pop:Double? = null,
 
 @SerializedName("sys")
 @Expose
